:truck: Feat : move location
